Reorganizing files to run from main
